Created FieldHolder HOC for TextField and HtmlReadonlyField
